// src/app/website/shared/models/user.model.ts
export class User {
    _id?: string;
    username?: string;
    email?: string;
    // role?: string;
    // fname?: string;
    // lname?: string;
    // dateofbirth?: string;
    // placeofbirth?: string;
    // timeofbirth?: string;
    // raasi?: string;
    // nakshatram?: string;
    // houseno?: string;
    // street?: string;
    // city?: string;
    // pin?: number;
    // state?: string;
    // profile_id? : string;
  }